@@ -1,11 +1,8 @@
 // Deprecated: go-mackerel-plugin-helper is still maintained, but we recommend using go-mackerel-plugin instead.
 module github.com/mackerelio/go-mackerel-plugin-helper
 
-go 1.17
+go 1.18
 
-require (
-	github.com/mackerelio/golib v1.2.1
-	golang.org/x/lint v0.0.0-20200302205851-738671d3881b
-)
+require github.com/mackerelio/golib v1.2.1
 
-require golang.org/x/tools v0.0.0-20200130002326-2f3ba24bd6e7 // indirect
+require golang.org/x/text v0.3.7
